+++ src/main/java/insane96mcp/insanesurvivaloverhaul/mixin/accessor/WorldGenRegionAccessor.java
@@ -1,0 +1,12 @@
+package insane96mcp.insanesurvivaloverhaul.mixin.accessor;
+
+import net.minecraft.server.level.WorldGenRegion;
+import net.minecraft.world.level.chunk.status.ChunkStep;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+@Mixin(WorldGenRegion.class)
+public interface WorldGenRegionAccessor {
+    @Accessor
+    ChunkStep getGeneratingStep();
+}